fix(fetch): Adapt Notifications to Anilist Login
Moved optimizedFetchAnimeOptions and optimizedFetchMangaOptions to dataFetch directory, removed optimizedFetchAnimeOptions.ts and optimizedFetchMangaOptions.ts files, updated imports in various components to reflect the changes.
