+++ app/watch/[id]/page.tsx
@@ -10,7 +10,7 @@ import CommentsSection from '@/app/components/CommentsSection'
 import aniwatch from '@/app/api/aniwatch'
 import VideoPlayer from './components/VideoPlayer'
 import { EpisodeAnimeWatch, EpisodeLinksAnimeWatch } from '@/app/ts/interfaces/apiAnimewatchInterface'
-import { optimizedFetchOnAniwatch, optimizedFetchOnGoGoAnime } from '@/app/lib/optimizedFetchAnimeOptions'
+import { optimizedFetchOnAniwatch, optimizedFetchOnGoGoAnime } from '@/app/lib/dataFetch/optimizedFetchAnimeOptions'
 import { ImdbEpisode, ImdbMediaInfo } from '@/app/ts/interfaces/apiImdbInterface'
 import { getMediaInfo } from '@/app/api/consumetImdb'
 import { SourceType } from '@/app/ts/interfaces/episodesSourceInterface'
